back/requirements.txt
Flask==3.0.3
redis==5.0.1
pyjwt==2.8.0
